tests seem to indicate success, as required.
