+++ api/tests/test_base.py
@@ -4,6 +4,8 @@ from rest_framework.test import APIClient
 from rest_framework import status
 from django.urls import reverse
 from django.conf import settings
+import random
+import string
 
 from api.models import (Ticket, Team, Member, Tag, TicketStatus, TicketTag, TicketNode)
 from ..serializers import UserSerializer, TicketStatusSerializer, TagSerializer, TicketSerializer, MemberSerializer, \
@@ -18,6 +20,60 @@ PAGE_SIZE = settings.REST_FRAMEWORK["PAGE_SIZE"]
 """
 Base class for all team related tests, querying CRUD operations
 """
+
+
+# make random count random users, while ensuring we don't have any duplicates
+def generateArbitraryUsers(count: int) -> list:
+    emails = set()
+    usernames = set()
+    users = []
+
+    for _ in range(count):
+        while True:
+            user = ''.join(random.choice(string.ascii_letters) for _ in range(10))
+            hostname = ''.join(random.choice(string.ascii_letters) for _ in range(10))
+            tld = ''.join(random.choice(string.ascii_letters) for _ in range(3))
+            email = f"{user}@{hostname}.{tld}"
+            if email not in emails:
+                emails.add(email)
+                break
+
+        while True:
+            username = ''.join(random.choice(string.ascii_letters) for _ in range(10))
+            if username not in usernames:
+                usernames.add(username)
+                break
+
+        first_name = ''.join(random.choice(string.ascii_letters) for _ in range(10))
+        last_name = ''.join(random.choice(string.ascii_letters) for _ in range(10))
+
+        users.append(User.objects.create(username=username,
+                                         email=email,
+                                         first_name=first_name,
+                                         last_name=last_name))
+
+        for user in users:
+            user.save()
+
+    return users
+
+
+def generateArbitraryTeams(count: int) -> list:
+    team_names = set()
+    teams = []
+
+    for _ in range(count):
+        while True:
+            team_name = ''.join(random.choice(string.ascii_letters) for _ in range(10))
+            if team_name not in team_names:
+                team_names.add(team_name)
+                break
+
+        team = Team.objects.create(name=team_name)
+        team.save()
+        teams.append(team)
+
+    return teams
 
 
 class TeamRelatedCore(TestCase):
