+++ api/urls.py
@@ -54,9 +54,6 @@ team_router.register(
 members_router = routers.NestedSimpleRouter(team_router, r'members', lookup='member')
 members_router.register(r'pinned_tickets', views.PinnedTicketViewSet, basename='pinned-tickets')
 
-user_router = SimpleRouter()
-user_router.register('invites', views.UserInviteView)
-
 urlpatterns = [
     # Optional UI:
     path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
@@ -67,7 +64,7 @@ urlpatterns = [
     path("api/", include(team_router.urls)),
     path("api/", include(members_router.urls)),
 
-    path("api/users/invites/", views.UserInviteView.as_view(), name="invites"),
+    path("api/users/invites/", views.UserInviteView.as_view(), name="user-invites"),
 
     path('auth/', include('dj_rest_auth.urls')),
     path('auth/registration/', include('dj_rest_auth.registration.urls')),
