+++ api/views/__init__.py
@@ -6,4 +6,5 @@ from .status import *
 from .member import *
 from .pinned_ticket import *
 from .team import *
-from .invite import *
+from .team_invite import *
+from .user_invite import *
